Files kept under LoginSystemDemo Folder
Simple Application Login System Demo

+++ LoginSystemDemo/admin/manage-users.php
@@ -7,11 +7,21 @@ if (strlen($_SESSION['adminid']==0)) {
 if(isset($_GET['id']))
 {
 $adminid=$_GET['id'];
-$msg=mysqli_query($con,"delete from users where id='$adminid'");
-if($msg)
-{
-echo "<script>alert('Data deleted');</script>";
+
+$validateDeletingUser=mysqli_query($con,"select * from users where id='$adminid'");
+$DoesUserExist=mysqli_num_rows($validateDeletingUser);
+if($DoesUserExist){
+	$msg=mysqli_query($con,"delete from users where id='$adminid'");
+	 
+	 if($msg)
+	{
+		echo "<script>alert('Data deleted');</script>";
+	}
 }
+else{
+	echo "<script>alert('User does not exist,please check the user before you delete it.');</script>";
+}	
+
 }
    ?>
 <!DOCTYPE html>
@@ -22,11 +32,11 @@ echo "<script>alert('Data deleted');</script>";
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
         <meta name="description" content="" />
         <meta name="author" content="" />
-        <title>Registered Users in Last 30 Days | Registration and Login System</title>
+        <title>Manage Users | Registration and Login System</title>
         <link href="https://cdn.jsdelivr.net/npm/simple-datatables@latest/dist/style.css" rel="stylesheet" />
         <link href="../css/styles.css" rel="stylesheet" />
         <script src="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/js/all.min.js" crossorigin="anonymous"></script>
-
+        
     </head>
     <body class="sb-nav-fixed">
       <?php include_once('includes/navbar.php');?>
@@ -35,16 +45,16 @@ echo "<script>alert('Data deleted');</script>";
             <div id="layoutSidenav_content">
                 <main>
                     <div class="container-fluid px-4">
-                        <h1 class="mt-4">Registered Users in Last 30 Days </h1>
+                        <h1 class="mt-4">Manage users</h1>
                         <ol class="breadcrumb mb-4">
                             <li class="breadcrumb-item"><a href="dashboard.php">Dashboard</a></li>
-                            <li class="breadcrumb-item active">Registered Users in Last 30 Days </li>
+                            <li class="breadcrumb-item active">Manage users</li>
                         </ol>
             
                         <div class="card mb-4">
                             <div class="card-header">
                                 <i class="fas fa-table me-1"></i>
-                          Registered Users in Last 30 Days  Details
+                                Registered User Details
                             </div>
                             <div class="card-body">
                                 <table id="datatablesSimple">
@@ -71,7 +81,7 @@ echo "<script>alert('Data deleted');</script>";
                                         </tr>
                                     </tfoot>
                                     <tbody>
-                                              <?php $ret=mysqli_query($con,"select * from users where date(posting_date)>=CURRENT_DATE()-30");
+                                              <?php $ret=mysqli_query($con,"select * from users");
                               $cnt=1;
                               while($row=mysqli_fetch_array($ret))
                               {?>
@@ -96,7 +106,7 @@ echo "<script>alert('Data deleted');</script>";
                         </div>
                     </div>
                 </main>
-  <?php include('../includes/footer.php');?>
+  
             </div>
         </div>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
